Raft integration (#38)
Introduces the raft protocol to the cluster for the purposes of leader election and high watermark consensus among signer nodes

All sign requests from the sentries made to any signer node will now be proxied through the raft leader.

The signer node that is the current elected raft leader will act upon the sign requests by managing the threshold validation process

The cluster can now reliably sign when (total signers - threshold signers) are down. If a signer node goes down and it is not the leader, there should not be any missed blocks. If the leader goes down, a new leader will be elected and it will pick up signing blocks after that, so there might be a few missed blocks in the case of a downed leader.